Regression with interaction
#Linear Regression
Step 1 You need to create the working directory and connect Boston data file as you did this in Home Work 1.
Step 2 include interaction terms in a linear model using the lm() function. The syntax lstat:black tells R to include an interaction term between lstat and black. The syntax lstat*age simultaneously includes lstat, age, and the interaction term lstat×age as predictors; it is a shorthand for lstat+age+lstat:age.
Step 3 Use command dim to check how many observations and variables are in the Boston file. Find the names of variables in this file.
Step 4 Run the multivariate regression of medv against lstat and age with interaction.
Step 5 Is interaction term significant or not. Is the answer difference for confidence probability 5% and 1%? Formulate the appropriate hypotheses and make a conclusion based on the relevant p-values.
Step 6 Implement the multiple linear regression of medv against lstat and age without interaction (or just see the results in HW1). Compare these two models with and without interaction.
Step 7 Compare the residual standard errors for these two models.

#Nonlinear transform of predictors
Step 1 The lm() function can also accommodate non-linear transformations of the predictors. For instance, given a predictor X, we can create a predictor X2 using I(X^2). The function I() is needed since the ^ has a special meaning I() in a formula; wrapping as we do allows the standard usage in R, which is to raise X to the power 2. We now perform a regression of medv onto lstat and lstat2.
Step 2 Build the single regression of medv (results can be taken from HW1) against lstat. Compare linear and quadratic model using the adjuster R-squared.
Step 3 Use the anova() function to further quantify the extent to which the quadratic fit is superior to the linear fit. The anova() function performs a hypothesis test comparing the two models. The null hypothesis is that the two models fit the data equally well, and the alternative hypothesis is that the full model is superior.

#Classification: Logistic regression
Step 1 Open library (ISLR) and check the name of variables in Smarket file and provide the summary statistics for all variables.
Step 2 Use the cor() function produces a matrix that contains all of the pairwise correlations among the predictors in a data set. Use the parameter [,9] because the Direction (#9) variable is qualitative.
Step 3 Explain why volume is correlated with the year. Illustrate this graphically.
Step 4 Fit a logistic regression model in order to predict Direction using Lag1 through Lag5 and Volume. The glm() function fits generalized linear models, a class of models that includes logistic regression. The syntax of the glm() function is similar to that of lm(), except that we must pass in the argument family=binomial in order to tell R to run a logistic regression rather than some other type of generalized linear model.
Step 5 Analyze the logistic Table in order our ability to predict trading Volume using the information about lags.
Step 6  Use the coef() function in order to access just the coefficients for this fitted model. Check the consistency with the previous step.
Step 7  Use the summary() function to access particular aspects of the fitted model, such as the p-values for the coefficients.
Step 8 Use the predict() function, which can be used to predict the probability that the market will go up, given values of the predictors. The type="response" option tells R to output probabilities of the form P(Y = 1|X), as opposed to other information such as the logit. If no data set is supplied to the predict() function, then the probabilities are computed for the training data that was used to fit the logistic regression model. Print the first ten probabilities.
Step 9 Use contrasts() function, which indicates that R has created a dummy variable with a 1 for Up, to check if these values correspond to the probability of the market going up, rather than down.
Step 10 Convert these predicted probabilities into class labels, Up or Down, in order to make our prediction more transparent. Check the result with command fix.
Step 11 Given these predictions, use the table() function to produce a confusion matrix in order to determine how many observations were correctly or incorrectly classified.
Step 12 Calculate the proportion of correct predictions using command mean. Check the results manually using the information from the confusion table.
Step 13 The statistics in Step 13 corresponds to the training set equal to the total set of observation. Create the new training set for records from 2001 to 2004. We will then use this vector to create a held out dataset of observations from 2005. How many records remain in the test set?
 Step 14 (1 mark) Implement command >Direction.2005= Direction [! train]